1/10
runes- initially it was going to be a lot more runes, but I didn't think I'd have time, and I suppose that's another story. These can be switched on or off with a bubble shot from the popper staff. When they're on, they glow, red blue or green , for 'no', 'maybe', and 'return'.  There's also a tribute to the game 'hyperrogue' which is like a school in hyperbolic geometry, but I've altered the logo to be without a sword, just the bull, and it's a base for a vertical road in the world 'Order', the minetest world file I made along with this minimal. There's also a set of rings which align to display the relative orbital periods of our solar system's inner planets.