popper effects flat_O itemholder, removed clipping on ladder lid
rubler updates with addition of stone models, horizontal lid subvolumes raised to avoid clipping striations. Note: the abm to remove empty entity prescense as 'ghost', was, through much trial and error implemented, but ghost remains a few seconds if item s are removed by staff projectile anyway, and is in sto3 with the flat_O nodereg, not po.